Changed platform and added string sending
First change is that the target is now the MEGA32 controller.
This is the one on the AVR stamp that Bertho has made available at
http://www.vagrearg.org/content/avrpicstamp

The second change is that by doing an arary of morse components
it is now possible to send an actual string.

I'm beginning to think we might need some kind of unit testing,
we've passed the 4k limit on this one, and since we have no debugger
it might be prudent to start testing to make sure we don't screw up.
Consider unity, since we are pure C (www.throwtheswitch.org)